crypto/requirements.txt
certifi==2017.11.5
chardet==3.0.4
cycler==0.10.0
idna==2.6
matplotlib==2.1.1
numpy==1.13.3
pyparsing==2.2.0
python-dateutil==2.6.1
pytz==2017.3
requests==2.18.4
six==1.11.0
urllib3==1.22
paho-mqtt==1.3.1